[level 3] Title: 다단계 칫솔 판매, Time: 35.52 ms, Memory: 140 MB -BaekjoonHub

+++ 프로그래머스/lv3/77486. 다단계 칫솔 판매/다단계 칫솔 판매.java
@@ -21,16 +21,16 @@ class Solution {
         return answer;
     }
 
-    private void shareProfit(double profit, String seller) {
+    private void shareProfit(int profit, String seller) {
         if (seller.equals("-")) {
             return;
         }
-        int nextProfit = (int) (profit / 10);
+        int nextProfit = profit / 10;
         if(nextProfit == 0){
-            profits.put(seller, profits.get(seller) + (int) profit);
+            profits.put(seller, profits.get(seller) + profit);
             return;
         }
-        profits.put(seller, profits.get(seller) + (int) (profit - nextProfit));
+        profits.put(seller, profits.get(seller) + (profit - nextProfit));
         shareProfit(nextProfit, referralMap.get(seller));
     }
 }
